vm: Add -Xtrace:compile command line option
The -Xtrace:jit command line option prints a lot of information for a
non-trivial programs as does -Xtrace:invoke which make them cumbersome
for determining which method caused a compilation error. Therefore,
introduce a -Xtrace:compile which only prints out methods when they are
compiled which produces significantly less output.

Signed-off-by: Pekka Enberg <penberg@cs.helsinki.fi>