vm: fix deadlock in find_class()
find_class() can be called by on a class which is being loaded by the
same thread which is loading the class. This happens when class
loading is delegated to external class loaders. We should handle this
by returning NULL so that external class loader will actually load the
class.

Signed-off-by: Tomek Grabiec <tgrabiec@gmail.com>
Signed-off-by: Pekka Enberg <penberg@cs.helsinki.fi>