vm: remove duplicate methods from itable
This fixes the lighthouse ticket:
http://jato.lighthouseapp.com/projects/29055-jato/tickets/4

Actually, the ticket describes a slightly different scenario. The scenario
from the ticket:

	interface A {
		void foo();
	}

	interface B {
		void foo();
	}

	class C implements A, B {
		void foo() { ... }
	}

Here, the itable for C will contain two entries, one for A.foo() and one
for B.foo(). However, we _must_ do this, since we use the _method address_
as the hidden parameter to the itable resolution stub. If we omitted one
of them, the resolution stub would not be able to find it, e.g.:

	C c = new C();

	((A) c).foo(); // would succeed
	((B) c).foo(); // would fail

What the ticket is meant to describe, however, is that we may get the same
method (from the same class) in the itable:

	interface I {
		void foo();
	}

	class A implements I {
		void foo() { ... }
	}

	class B extends A implements I {
		void foo() { ... }
	}

Here, before this patch, we would add I.foo() twice to the itable of B.

Signed-off-by: Vegard Nossum <vegard.nossum@gmail.com>
Signed-off-by: Pekka Enberg <penberg@cs.helsinki.fi>