jit: fix conversion of invoke* instructions
It is incorrect to convert invocations into expressions which are
pushed onto mimic stack because pushed expressions can be evaluated
any time later after they are pushed. This can result in breaking the
execution sequence, where methods are not invoked in the order they
appear in bytecode.

Invocations are now handled as statements and their result is pushed
onto mimic stack.

Signed-off-by: Tomek Grabiec <tgrabiec@gmail.com>
Signed-off-by: Pekka Enberg <penberg@cs.helsinki.fi>