gc: Pass register state to gc_start() and gc_safepoint()
Save registers in gc_alloc() and pass them to gc_safepoint() so the
latter can scan them for references. Also save the registers in signal
handler and pass them to gc_safepoint(). We need to use pre-allocated
thread-local memory because we can't use the stack or malloc() in the
signal handler.

Cc: Tomek Grabiec <tgrabiec@gmail.com>
Cc: Vegard Nossum <vegard.nossum@gmail.com>
Acked-by: Arthur HUILLET <arthur.huillet@free.fr>
Signed-off-by: Pekka Enberg <penberg@cs.helsinki.fi>