x86: Register state saving
This patch introduces a save_registers() function that can be used to
take a snapshot of general purpose registers. This is needed in the
gc_alloc() function to save registers in case we need to enter a
safepoint and need to scan them for references.

We also need a save_signal_registers() function for converting gregset_t
to struct register_state so we can pass register contents to the
gc_safepoint() function.

Cc: Tomek Grabiec <tgrabiec@gmail.com>
Cc: Vegard Nossum <vegard.nossum@gmail.com>
Acked-by: Arthur HUILLET <arthur.huillet@free.fr>
Signed-off-by: Pekka Enberg <penberg@cs.helsinki.fi>