x86: introduce valgrind workaround for exception guards.
Installing signal bottom half doesn't work well in valgrind
environment (it causes an error and application is killed). This
introduces a workaround for this situation when caused by exception
guards. Exception guards are used to catch signalled exceptions via
SIGSEGV handler and transfer control to the exception handler or
unwind. The workaround is to inline exception check and don't use
signal handlers when run on valgrind.

Signed-off-by: Tomek Grabiec <tgrabiec@gmail.com>
Signed-off-by: Pekka Enberg <penberg@cs.helsinki.fi>