jit: fix bug in spill_interval()
The function did not handle properly the case when we are splitting
and interval which has child intervals that need reloading from it.
We should change the spill parent of those child intervals to the
new interval, so that they are reloaded correctly.

This bug caused execution of invalid methods due to register
corruption, which prevented java.awt.image.Raster.createWritableRaster()
from correct working.

Signed-off-by: Tomek Grabiec <tgrabiec@gmail.com>
Signed-off-by: Pekka Enberg <penberg@cs.helsinki.fi>