jit: skip conversion of unreachable basic blocks
Basic blocks which are unreachable in CFG and are not
exception handlers will not be converted.

This is a workaround for a problem encountered in bytecode generated
by clojure. Clojure generates unreachable basic blocks which leave
empty mimic stack and jump to a basic block which expects a value on
mimic stack. This breaks mimic stack resolution because we have
different stack depths on different paths (bb1->bb3 and bb2->bb3):

  -- bb1 (unreachable)
  [main]    [ 354 ]  0xa7  goto           359

  -- bb2
  [main]    [ 357 ]  0x57  pop
  [main]    [ 358 ]  0x01  aconst_null

  -- bb3
  [main]    [ 359 ]  0xb0  areturn

Signed-off-by: Tomek Grabiec <tgrabiec@gmail.com>
Signed-off-by: Pekka Enberg <penberg@cs.helsinki.fi>