vm: fix loading of array classes.
Array classes have classloader set to the classloader of its
element class. For primitive arrays, the classloader is always
bootstrap classloader.

This is a bug fix which reveals itself after introducing a
classloader-aware cache. Loading of array classes [I and [[I can be
initiated with different classloaders. The former is always loaded
with bootstrap classloader (primitive array) while loading of the
latter can be initiated with any classloader. Class [[I must be
eventually loaded with bootstrap classloader too.

Signed-off-by: Tomek Grabiec <tgrabiec@gmail.com>
Signed-off-by: Pekka Enberg <penberg@cs.helsinki.fi>