jit: set bytecode offset of spill/reload instructions properly
Functions get_spill_after_insn() and get_reload_before_insn() not
always return a pointer to struct insn (the API will be fixed in
following patches). This can happen when we must insert spill
instruction at the beginning of basic block, or a reload instruction
at the end of basic block. In those cases a container of insn list
head is returned. Therefore we can't get the bytecode offset by
dereferencing the result. A valid bytecode offset value is returned by
those function by setting a variable to which pointer is passed as an
argument.

Signed-off-by: Tomek Grabiec <tgrabiec@gmail.com>
Signed-off-by: Pekka Enberg <penberg@cs.helsinki.fi>