vm: fix init_stack_trace_elem_current()
The function was setting elem to the init_stack_trace_elem_current()'s
frame which is error prone because after function returns its stack
frame might be overriden by another function calls done from the
caller.

A fix for that is to make init_stack_trace_elem_current() set elem to
the caller's frame. We must make it a macro becasue we can't rely on
__builtin_frame_address(1) because of gcc optimizations.

Signed-off-by: Tomek Grabiec <tgrabiec@gmail.com>
Signed-off-by: Pekka Enberg <penberg@cs.helsinki.fi>