x86: Unify native_call_eax() and native_call_long()
This patch unifies the native_call_eax() and native_call_long() functions. The
only difference between them is that the latter handles 64-bit return values
whereas the former ignores the upper 32 bits.

That's pointless because the 32-bit callers only look at the bottom 32 bits and
there's no performance gain because we're touching the same cacheline for the
upper 32 bits.

Cc: Tomek Grabiec <tgrabiec@gmail.com>
Cc: Eduard - Gabriel Munteanu <eduard.munteanu@linux360.ro>
Signed-off-by: Pekka Enberg <penberg@cs.helsinki.fi>