x86: Generic instruction encoding
This patch introduces insn_encode() function that's a generic x86 table-driven
instrunction encoder. The intent behind the patch is to introduce framework for
converting all instrunction encoder to the new approach to make the code more
maintainable and less buggy because we can consolidate all the different ModR/M
and SIB byte encoders.

Cc: Arthur HUILLET <arthur.huillet@free.fr>
Cc: Eduard - Gabriel Munteanu <eduard.munteanu@linux360.ro>
Cc: Tomek Grabiec <tgrabiec@gmail.com>
Signed-off-by: Pekka Enberg <penberg@cs.helsinki.fi>