jit: fix vtable fixup
After method compilation the vtable fixup was done for the class in
which the method was declared, not the actual class of the
object. This caused that invokevirtual of
java/util/concurrent/locks/AbstractQueuedSynchronizer.isOnSyncQueue in
java/util/concurrent/locks/AbstractQueuedSynchronizer$ConditionObject.await()
(with 'this' referencing object inheriting AbstractQueuedSynchronizer)
was always calling a jit_magic_trampoline() which is quite expensive.

This fix is a sort of revert of commit c0c9caa1c6a8c5f40fc1b4bb5026fc07e2a7ca2a.
It seems that the way vtable is constructed changed since that commit
and it is no longer invalid to fix vtable of the actual object's class.

Signed-off-by: Tomek Grabiec <tgrabiec@gmail.com>
Signed-off-by: Pekka Enberg <penberg@cs.helsinki.fi>