x86: Optimize fixup_direct_calls() locking
This patch optimizes fixup_direct_calls() not to take heavy pthread mutex in
case the fixup list is empty. It's possible that we miss a fixup being added to
the list concurrently but we don't care because that thread will eventually do
the fixup for us anyway.

This speeds up DaCapo antlr benchmark by 13%:

  ===== DaCapo antlr PASSED in 19265 msec =====

  real    0m22.223s
  user    0m19.113s
  sys     0m0.604s

  ===== DaCapo antlr PASSED in 16651 msec =====

  real    0m19.213s
  user    0m16.598s
  sys     0m0.580s

Acked-by: Tomek Grabiec <tgrabiec@gmail.com>
Signed-off-by: Pekka Enberg <penberg@cs.helsinki.fi>