vm: introduce -Xtrace:gate <regex>
When -Xtrace:gate is enabled then all invocation traces
are generated only for invocations which are enclosed in
invocations of methods matching the gate method regex.

It is useful when we want to get full invocation trace
only on some execution paths so that we are not flooded by
useless data.

Signed-off-by: Tomek Grabiec <tgrabiec@gmail.com>
Signed-off-by: Pekka Enberg <penberg@cs.helsinki.fi>