vm: fix bug in assuring .object of struct vm_class
.object was swapped in vm_class_ensure_object() which led to
SIGSEGVs because threads could be waiting on monitor when
object was swapped.

The solution is to ensure .object immediately after classes
are preloaded. During preload .object points to a dummy structure
for class initialization synchronization to work.

We can not allocate propper object for classes until preload is
finished because allocation of object requires vm_java_lang_Class
to be initialized and vm_java_lang_Class_vmdata field to be
preloaded.

Signed-off-by: Tomek Grabiec <tgrabiec@gmail.com>
Signed-off-by: Pekka Enberg <penberg@cs.helsinki.fi>