jit: use light weight locking in trampoline
A new kind of locking mechanism is introduced called compile lock
which can be used to synchronize compilation and initialization
processes.

The lock has an associated status: INITIAL, COMPILING, COMPILED_OK
and COMPILED_ERRONOUS. Only one thread is permited to enter compilation
and update the status. Other threads trying to enter compilation will
be blocked until compilation is finished.

If state is greater than COMPILING then it can not change. When
compilation data is not updated after compilation, and we know the
status is greater than COMPILING then this data can be accessed
without any additional locking safely.

Signed-off-by: Tomek Grabiec <tgrabiec@gmail.com>
Signed-off-by: Pekka Enberg <penberg@cs.helsinki.fi>