jit: fixup vtable entry for the class of this when in invokevirtual
After method compilation the vtable fixup was done for the class in
which the method was declared, not the actual class of the
object. This caused that invokevirtual of
java/util/concurrent/locks/AbstractQueuedSynchronizer.isOnSyncQueue in
java/util/concurrent/locks/AbstractQueuedSynchronizer$ConditionObject.await()
(with 'this' referencing object inheriting AbstractQueuedSynchronizer)
was always calling a jit_magic_trampoline() which is quite expensive.

Signed-off-by: Tomek Grabiec <tgrabiec@gmail.com>
Signed-off-by: Pekka Enberg <penberg@cs.helsinki.fi>