vm: optimize stack trace filling
There are applications which heavily rely on exceptions for
control flow. In most cases applications do not need the
stack trace. The runtime provides support for this
by allowing a VM to create an intermediate stack trace
when throwable object is created. This intermediate
stack trace is converted to full stack trace only when
getStackTrace() on throwable is called.

This patch optimizes intermediate stack trace filling
by not doing a compilation unit lookup, which is expensive.
Only native address is stored and lookup is done in
convert_intermediate_stack_trace().

This gives 7% speedup in dacapo lusearch.

Signed-off-by: Tomek Grabiec <tgrabiec@gmail.com>
Signed-off-by: Pekka Enberg <penberg@cs.helsinki.fi>