vm: preliminary support for weak, soft and phantom references
The API can be used both by JIT and VM code. The JIT code
creates weaker references through java.lang.ref.Reference class.
putfield and getfield on .referent field of that class are
redirected to put_referent() and get_referent() functions.
put_referent() creates struct vm_reverence and puts pointer to it
in the .referent field. We do not want to store pointer
to referenced object directly because boehm GC would see that
pointer as a strong reference.

When vm_reference is created for some object then it is
forced that vm_object_finalizer() will be called for that object.

Just before object is finalized, all weaker references to it are cleared.
If reference was created for java.lang.ref.Reference then clear() and
enqueue() methods are called for it.

Signed-off-by: Tomek Grabiec <tgrabiec@gmail.com>
Signed-off-by: Pekka Enberg <penberg@cs.helsinki.fi>