x86: Fix bottom half FPU state clobbering
Due to static (class) initialisers, floating-point code can be executed without
saving and restoring the floating-point registers. Take for example this code:

[main]   0xe8074dc1:   f2 0f 10 7d f8             movsd  -0x8(%ebp),%xmm7
[main]   0xe8074dc6:   f2 0f 10 35 00 50 1e 08    movsd  0x81e5000,%xmm6
[main]   0xe8074dce:   f2 0f 10 f6                movsd  %xmm6,%xmm6

Here, the second instruction, movsd  0x81e5000,%xmm6, will actually touch a non-
present page and cause the initialisation of a static inner class. Here is part
of the code for the initialiser:

[main]   0xe8074e70:   c7 45 f4 00 00 00 00       movl   $0x0,-0xc(%ebp)
[main]   0xe8074e77:   f2 0f 10 7d f4             movsd  -0xc(%ebp),%xmm7
[main]   0xe8074e7c:   f2 0f 11 3d e0 6a 21 08    movsd  %xmm7,0x8216ae0

It is clear that its use of %xmm7 interferes with the use in the first method
when these registers are not saved or restored anywhere.

This patch fixes the problem by adding FPU state saving and restoring to
signal_bh_trampoline().

Signed-off-by: Vegard Nossum <vegard.nossum@gmail.com>
Signed-off-by: Pekka Enberg <penberg@cs.helsinki.fi>