vm: fix array class loading
Array classes are always created by VM but the defining classloader for
them is set to the defining class loader of array element. This not necessarily
mean that this will be the class loader which initiated loading, as was done
before this patch.

For example if the system class loader loads [Ljava/lang/Object; then
the defining class loader for this class should be set to NULL, because
java/lang/Object is _defined_ by the bootstrap classloader.

This unblocks some dacapo-9.04-bach benchmarks.

Signed-off-by: Tomek Grabiec <tgrabiec@gmail.com>
Signed-off-by: Pekka Enberg <penberg@cs.helsinki.fi>