jit: allow compilation unit recompilation in jit_magic_trampoline
If method compilation fails we should not mark it as indefinitely
erronous, like for java classes, because the conditions which make
method compilation failing may change. One example of this is
compilation failure due to UnsatisfiedLinkError. The following
code from equinox reacts on the error and triggers re-compilation
of native method _update_splash():

public boolean updateSplash() {
		try {
			_update_splash();
			return true;
		} catch (UnsatisfiedLinkError e) {
			if(!libraryLoaded){
				loadLibrary();
				return updateSplash();
			}
			return false;
		}
	}

After call to loadLibrary() the native library is loaded
and the native for _update_splash() can be resolved.

Signed-off-by: Tomek Grabiec <tgrabiec@gmail.com>
Signed-off-by: Pekka Enberg <penberg@cs.helsinki.fi>