jit: Dominance frontier computation for SSA form
This patch implements the algorithm described in the paper "A Simple, Fast
Dominance Algorithm" (2001) by Cooper et al. I also used Mono as a reference
which explains some of the differences between our implementation and the
paper.

Exception handler basic blocks are completely ignored which might cause
problems in the SSA form conversion.

Cc: Arthur HUILLET <arthur.huillet@free.fr>
Cc: Tomek Grabiec <tgrabiec@gmail.com>
Signed-off-by: Pekka Enberg <penberg@kernel.org>