x86: Separate 32-bit and 64-bit instruction selectors
Keeping 32-bit and 64-bit instruction selectors in the same file seemed like a
good idea. However, the current ifdef maze is pretty much unreadable and
actually doesn't help things on the 64-bit port.

It's pretty evident by now that 64-bit instruction selection is going to be
very different because registers are used to pass arguments to functions and
because 32-bit requires a register pair for 64-bit operations. Therefore, split
the two instruction selectors.

Cc: Eduard - Gabriel Munteanu <eduard.munteanu@linux360.ro>
Signed-off-by: Pekka Enberg <penberg@kernel.org>