jit: Fix convert_tableswitch and convert_lookupswitch
This patch fixes a bug in the generation of basic block
successors of a basic block containing the
lookupswitch_jump to target instruction (for lookupswitch)
or the tableswitch instruction (for tableswitch).

The problem was that the basic block immediately following the 2
basic blocks mentioned before was duplicated in their basic block
successors list. This was because the basic block was put in the
successor list twice (once in split_after_branches and once in
convert_lookupswitch or convert_tableswitch).

The problem is solved by checking if the basic block isn't
already in the basic block successor list when inserting it
in convert_lookupswitch and convert_tableswitch.

Cc: Eduard - Gabriel Munteanu <eduard.munteanu@linux360.ro>
Signed-off-by: Ana Farcasi <farcasia@gmail.com>
Signed-off-by: Pekka Enberg <penberg@kernel.org>