jit: Support for multiple entry points to methods.
This patch adds support for adding multiple entry points
to methods. The entry points are stored in *_entry_point
fields of the struct compilation_unit. cu->entry_point
stores the normal entry point to a method. cu->ic_entry_point
will hold the entry point to be used by call sites with an
inline cache.

The functions cu_entry_point() or cu_ic_entry_point() should be
used for getting a pointer to call the method. These functions
replace the call buffer_ptr(cu->objcode) for getting the entry
point for a method.

buffer_ptr(cu->objcode) should be used for getting the starting
address of the native code buffer of the method.

Signed-off-by: Balagopal <baluks@gmail.com>
Signed-off-by: Pekka Enberg <penberg@kernel.org>