jit: Added renaming of variables for exception handler basic blocks
This patch renames virtual registers in exception handler basic blocks. This is required
in order to have unique vreg values for all var_info structures. If virtual registers in
exception handler basic blocks would not be renamed, then virtual registers with the
same vreg value may exist and compromise the results of the register allocator.

Signed-off-by: Ana Farcasi <farcasia@gmail.com>
Signed-off-by: Pekka Enberg <penberg@kernel.org>