jit: Fix bug for insn_use_def in ssa.c
This patch fixes a bug in ssa.c. Till now, we just verified if the
instruction uses and defines the same register, and we kept in
insn_add_ons the variable from the register. But, this verification is
not enough, having to check whether the source register or whether
the destination register is used and defined.

Signed-off-by: Ana Farcasi <farcasia@gmail.com>
Signed-off-by: Pekka Enberg <penberg@kernel.org>