Makefile: Framework to generate asm-offsets
This patch adds a framework to generate arch/asm-offsets.h
which contains macro definitions to be used in asm
sources.

Each arch/$(ARCH) folder contains a file "asm-offsets.c" which
will be compiled and run to generate the definitions. The
definitions must be printed to stdout. If "asm-offsets.c" is
not present, an empty "asm-offsets.h" is generated.

Signed-off-by: Balagopal <baluks@gmail.com>
[ penberg@kernel.org: cleanups and .gitignore ]
Signed-off-by: Pekka Enberg <penberg@kernel.org>