jit: Fix bb successors list for tableswitch
This patch fixes a bug in jit/switch_bc.c. This bug has been discovered
compiling the following method:

public int conditional()
{
        int a = 1;

        switch(a) {
                case 0:
                        a = 10;
                case 1:
                        a = 30;
                case 5:
                        a = 40;
        }

        return a;
}

Tableswitch introduces the same basic block twice in the list of basic
block successors because the same basic block appears in two case
branches:

[main] TABLESWITCH:
[main]   index: [temporary int 0xa1122a0 (low)]
[main]   case 0: [bb 0xa113020]
[main]   case 1: [bb 0xa11ef90]
[main]   case 2: [bb 0xa121dd0]
[main]   case 3: [bb 0xa121dd0]
[main]   case 4: [bb 0xa11f0b0]

This patch fixes the problem by introducing a basic block in the
successors list only if it is not already present in this list.

Signed-off-by: Ana Farcasi <farcasia@gmail.com>
Signed-off-by: Pekka Enberg <penberg@kernel.org>