jit: All virtual regs are defined before they are used
This patch adds an initialization instruction in the entry basic block
for those virtual registers that are used before they are defined. This
can happen when a register is defined before it is used in the dominator
basic block order, but it is being used before it is being defined in
the normal basic block order (the one imposed by the bytecode offset).
It is necessary that a virtual register is defined before it is used
in the normal basic block order for the register allocator.

Signed-off-by: Ana Farcasi <farcasia@gmail.com>
Signed-off-by: Pekka Enberg <penberg@kernel.org>