jit: Fix bug in resolve data flow
There are situations when the register allocator discovers useless
intervals and doesn't allocate any physical registers to them. These
intervals can be situated at the end of the normal basic block flow
order, order imposed by the bytecode offset. This placement of useless
intervals can happen when the flow order of basic blocks dictated by
the dominator tree is differrent from the normal one. If the useless
interval stretches entire basic blocks, it means that those basic
blocks don't have a use position for the virtual register. That is why
resolving is not necessary for destination basic blocks or source
basic blocks that don't have a real destination interval.

Signed-off-by: Ana Farcasi <farcasia@gmail.com>
Signed-off-by: Pekka Enberg <penberg@kernel.org>