arm: conversion of EXPR_VALUE to LIR
This commit is related to the conversion of EXPR_VALUE HIR to LIR
arm doesn't support immediates more than 8 bit so i have implemented
constant literal pool for immediates more than 8 bit.
8 bit or less immediates rae converted directly into reg_imm instruction
and immediates more than 8 bit are stored in constant literal pool which
will be emitted at the end of each method(compilation unit).
Pool is implemented using a singly linked list.
First compiler will serach for the immediate value in the pool if its there
then it will not create a new node for that imm value otherwise it
will create a new node for that immdiate value.
see constant-pool.c for details.

Test cases are also included for LONG as well as INT immediate values.

Signed-off-by: Ankit Laddha <ankit1991laddha@gmail.com>
Signed-off-by: Pekka Enberg <penberg@kernel.org>