x86: Fix bug in insert_copy_slot_64_insns
This patch fixes a bug in insert_copy_slot_64_insns. This
bug produced a crash because it was trying to get the second
slot for a 64 bit variable from the local slots. This function
is used to insert spill and reload instructions and needs
slots from frame->spill_slots and not from frame->local_slots.
In the future, if this function will need slots from the
local slots too, it can be modified.

Signed-off-by: Ana Farcasi <farcasia@gmail.com>
Signed-off-by: Pekka Enberg <penberg@kernel.org>