jit: Replace minimal SSA with pruned SSA
This patch replaces minimal SSA with pruned SSA. Pruned SSA is
that form of SSA that inserts the least number of phi instructions,
because it inserts phi instructions only for those variables that
are in the Live In set of the basic block that is in the dominance
frontier set of another basic block in which the variable is defined.
This also means that we don't need to insert initialization instructions
into the entry basic block as before, because a correct instruction
flux only uses a non-fixed variable only if it has been defined and
pruned SSA does not break this rule.

The time results obtained before this patch are:

ana@ubuntu:~/gsoc/jato$ sudo perf stat -e cpu-clock -r 10
	./jato -Xssa -cp test/functional jvm/EntryTest

 Performance counter stats for
	'./jato -Xssa -cp test/functional jvm/EntryTest' (10 runs):

       8072.537255 cpu-clock-msecs            ( +-   0.147% )

        8.109375133  seconds time elapsed   ( +-   0.195% )

The time results obtained after this patch are:

ana@ubuntu:~/gsoc/jato$ sudo perf stat -e cpu-clock -r 10
	./jato -Xssa -cp test/functional jvm/EntryTest

 Performance counter stats for
	'./jato -Xssa -cp test/functional jvm/EntryTest' (10 runs):

        975.172618 cpu-clock-msecs            ( +-   0.547% )

        0.987164466  seconds time elapsed   ( +-   1.086% )

Signed-off-by: Ana Farcasi <farcasia@gmail.com>
Signed-off-by: Pekka Enberg <penberg@kernel.org>