jit: Fix bug in control flow graph
This patch fixes a bug in the generation of the control flow
graph, bug that produced a crash when running jato with SSA
enabled because SSA relies heavily on the control flow graph
of the program being compiled.

The bug consisted in adding the same successor basic block
twice in the list of successors basic blocks of a basic block.
The successor was added first after branch splits (it was the
next basic block B after a basic block A ending with a conditional
jump) and the second time was added because it was the target
basic block of the first basic block (A).

This bug was discovered running jato with SSA enabled on dacapo
and appeared in antlr/JavaCodeGenerator.gen.

Signed-off-by: Ana Farcasi <farcasia@gmail.com>
Signed-off-by: Pekka Enberg <penberg@kernel.org>