arm: Fix the flow of process of encoding insns
Earlier we were first ending the data processing(DP) insns
and then the load-store insn but when using if it was
causing problem beacause DP insns have 0b00 as flag.
now we are encoding DP insns at last.

this bug is first noticed when i tried to encode
mov_reg_reg_insn.

Signed-off-by: Ankit Laddha <ankit1991laddha@gmail.com>
Signed-off-by: Pekka Enberg <penberg@kernel.org>