arm: Encode INSN_UNCOND_BRANCH
Signed-off-by: Ankit Laddha <ankit1991laddha@gmail.com>

This commit Encodes the INSN_UNCOND_BRANCH lir.
The most important work in the process of encoding is to determine
the branch address.We use PC relative addressing as the addressing
mode.
We backpatch those insns for which we can't find out the branch
address yet and afterwards emit them.
Signed-off-by: Pekka Enberg <penberg@kernel.org>