jit: Improve traversal of a bitset
This patch improves the way set bits are searched in a bitset.
The search starts from a given index in the bitset.

Until this patch, ssa was using bitset_ffs to traverse a bitset.
Perf profiles were indicating that bitset_ffs was creating much overhead.
>From now on, ssa uses bitset_ffs_from to search for set bits in a
bitset.

This patch improves the time with only a few hundreths of
seconds.

Signed-off-by: Ana Farcasi <farcasia@gmail.com>
Signed-off-by: Pekka Enberg <penberg@kernel.org>