vm, verifier: Almost completely implementing the first verification pass.
The first phase of verification (as explicited in section 4.9.2, before
the data-flow analysis) is nearly completed. Are missing the analysis of
the offsets in the tableswitch and lookupswitch instructions, the
verification of the exception handlers limit and functionnal tests to
assert everything works fine (already hand-tested to some extent, and
passes make check).

Signed-off-by: Theo Dzierzbicki <theo.dz@gmail.com>
[ penberg@kernel.org: fix make clean ]
Signed-off-by: Pekka Enberg <penberg@kernel.org>