arm: Emission of trampoline started
This commit emits the instructions for trampoline till
the calling of magic_trampoline.
Now we can call jit_magic_trampoline from trampoline.

We first emit the constant pool containing the adrresses
of compilation unit and magic trampoline because we have
only 8 bit immediates on ARM.
After that we emit the insns to load the address of cu in
R0 and to call the magic_trampoline.

Signed-off-by: Ankit Laddha <ankit1991laddha@gmail.com>
Signed-off-by: Pekka Enberg <penberg@kernel.org>