arm: Fix bug in trampoline emission
The call to jit_magic_trampoline() is emiited by two isns
First we calculate LR and then we mov the address to pc.
Earlier when calculating LR we are storing the address of
mov insn rather than then storing the address of next insn
from mov.That's why we were ending up calling
jit_magic_trampoline() in an infinite loop.

now this bug is fixed by re calculating the value of LR.

Signed-off-by: Ankit Laddha <ankit1991laddha@gmail.com>
Signed-off-by: Pekka Enberg <penberg@kernel.org>