jit: Fix bb successors list for lookupswitch
This patch fixes a bug in jit/switch-bc.c. This bug has been discovered
runing dacapo bloat benchmark. Lookupswitch was introducing the same
basic block multiple times in the successors basic block list of the
basic block containing the lookupswitch_jump instruction.

This patch fixes the problem by introducing a basic block in the
successors list only if it is not already present in this list.

Signed-off-by: Ana Farcasi <farcasia@gmail.com>
Signed-off-by: Pekka Enberg <penberg@kernel.org>