test/jit: Fix bug in switch-bc-test
This test fixes some bugs in switch-bc-test.c.

Until this patch, the testing methods in switch-bc-test.c
were starting with test1_ and test2_ which resulted in them
not being run because make-tests.sh script was looking for
methods starting with "test_" in order to be taken in
account.

This patch renames the methods and fixes a bug in one of them.

Signed-off-by: Ana Farcasi <farcasia@gmail.com>
Signed-off-by: Pekka Enberg <penberg@kernel.org>