vm, verifier: starting CFG analysis, proper tracing
First shot at implementing the CFG analysis functions (calls commented as
not working yet). Advanced type checking (taking into consideration
user-defined classes and interfaces) needed ASAP to get further.

Added a "-Xtrace:verifier" option to Jato and limited output of the
verifier to a minimum when it is not set.

Signed-off-by: Theo Dzierzbicki <theo.dz@gmail.com>
Signed-off-by: Pekka Enberg <penberg@kernel.org>