jit: Compute natural loops
This patch computes the natural loops from the control flow
graph of the program. The natural loop of a back edge is the
smallest set of nodes that includes the head and tail of the
back edge, and has no predecessors outside the set, except
for the predecessors of the header.

Natural loop detection is done in order to facilitate the
implementation of optimizations on loops.

This patch has been implemented based on the implementation
of mono mini's compiler and on the definition of the natural
loop.

Signed-off-by: Ana Farcasi <farcasia@gmail.com>
Signed-off-by: Pekka Enberg <penberg@kernel.org>