lib: Implement new ZIP handling for bootstrap classloader
This patch implements light-weight ZIP handling to be used in the boostrap
classloader similar to what CACAO does. The main advantage over libzip is that
we construct a hash map of class names during zip file loading for fast
lookups. Our implementation also uses mmap() for reading the zip files which is
faster than the buffered I/O libzip uses.

Signed-off-by: Pekka Enberg <penberg@kernel.org>