x86, x86-64: restore the stack frame correctly when unwinding
In addition to space reserved for locals, unwinding has to account for
callee-saved registers and other values saved on the stack. Both x86 and
x86-64 were affected, the former since it adopted a few stack layout
changes (standard prologs/epilogs and canaries), while the latter seems
to have been affected for a long time.

Signed-off-by: Eduard - Gabriel Munteanu <eduard.munteanu@linux360.ro>
Signed-off-by: Pekka Enberg <penberg@kernel.org>